+++ src/space.h
@@ -8,8 +8,8 @@ extern CFArrayRef SLSCopyWindowsWithOptionsAndTags(int cid, uint32_t owner, CFAr
 
 CFStringRef space_display_uuid(uint64_t sid);
 uint32_t space_display_id(uint64_t sid);
-uint32_t *space_window_list_for_connection(uint64_t sid, int cid, int *count);
-uint32_t *space_window_list(uint64_t sid, int *count);
+uint32_t *space_window_list_for_connection(uint64_t sid, int cid, int *count, bool include_minimized);
+uint32_t *space_window_list(uint64_t sid, int *count, bool include_minimized);
 CFStringRef space_uuid(uint64_t sid);
 int space_type(uint64_t sid);
 bool space_is_user(uint64_t sid);